Fix split tool button drawing on Vista when not in a tool bar
When a tool button is not in a tool bar on XP and Vista it will get a
slightly different appearance from normal tool buttons. I resolved
this by drawing a normal tool button with a divider line on top
if the autoraise property is not set on the button. (which is by
default enabled only for buttons in a tool bar).

Task-number: QTBUG-5061
Reviewed-by: prasanth