(Windows) Don't create a file mapping for each view that is mapped
Creating a file mapping for each view mapped to memory is sub-optimal
and slow. With this change, a single mapping is created and reused when
mapping subsequent views of the file.

The handle returned from CreateFileForMapping (used in WinCE 5) is now
discarded, since the kernel manages it automatically with the file
mapping. This simplifies use of the deprecated map API.

Reviewed-by: Maurice Kalinowski
Reviewed-by: Marius Storm-Olsen