Preventing unnecessary calls to CVideoPlayerUtility::SetDisplayWindowL
getNativeWindowSystemHandles now checks the window handle and screen
rectangle against the previous value; SetDisplayWindowL is only called
if the window and/or screen rectangle have changed.  This allows
videoOutputRegionChanged to be called 'speculatively' - i.e. in response
to Qt events which may or may not reflect a change in the underlying
window system - with only actual window system events getting propagated
into the MMF.

The reason for this change is that SetDisplayWindowL results in the
current DSA session (owned by CVideoPlayerUtility) being torn down and
a new one set up.  This in turn requires handshaking with the window
server, and may be slow.

Reviewed-by: Frans Englich