Added a new event type, WinIdChange.
This is sent to a native widget when its window system identifer has
changed.

This is motivated by the fact that, on Symbian, the native window
system identifier may change in situations other than a change of
parent widget.  Specifically, calling QWidget::setParent, passing in
the widget's existing parent, but OR-in Qt::Window into the window
flags, causes a new native window handle to be created.  Furthermore,
because of the fact that Symbian does not allow existing windows to be
reparented, any descendents of the original widget which are also
native, must also be given new window system handles.

Note that setWinId does not send a WinIdChange event if the incoming
winId is zero.  This is because setWinId(0) is only called in two
situations:
1. During native widget destruction
2. During re-creation of the winId for a native widget

Task-number: QTBUG-4664
Reviewed-by: Bjoern Erik Nilsen