Stub out flush() in the PowerVR GL window surface class
When QGLWidget was used as a viewport for QGraphicsView on PowerVR/MBX
systems, it was double-flushing every frame because the window surface
flush() implementation was still trying to do a raster blit after painting.
This change suppresses the raster blit, leaving it up to the GL
swapBuffers() call to get the painted contents onto the screen.

Reviewed-by: Sarah Smith