Ensure that qmake doesn't lose the error code when processing subdirs
When processing a project in a subdirs template failed for whatever
reason then qmake would lose the result of that and would return an
error code of 0 if the subdirs project file itself was processed fine.

So now it ensures that any errors arising from processing a project
referenced in a subdirs project file are not lost so that the error code
returned from qmake will indicate an error actually occured.

Task-number: QTBUG-4065
Reviewed-by: mariusSO
Original-commit: c15b370c9db16fdbfd9e7bec89ee9bf8c1110827