DialogButtonBox is shown together with softkeys (QTBUG-4739)
This fixes issue that both DialogButtonBox and softkeys are visible
(e.g. in Ftp example) at the same time.
Reason is that the previous implementation did the visiblitity setting when buttons were created.
Unfortunately buttons can be created without parent and let layout manager to find parent them
for you - so dialog as parent (=hide dialogbuttonbox as softkeys are in use) check failed.
Now similar check is done before showing the widget - it is layoutted here, so it has parent as well.
As a bonus, the event handling is already looking for dialog parent, so we can avoid doing another
for-loop.

Task-number: QTBUG-4739