Fix context menu and menu key in QGraphicsView.
This fixes a weird bug when pressing the menu key on a proxy that
embed a lineedit (but the proxy don't have the focus).
At each time the key was pressed a context menu was created. When we press
the menu key, QWidgetMapper will create a context menu event that will be
sent to the QGraphicsView and will end up in QGraphicsScene.
QGraphicsScene will then try to find all items under the mouse and will
try to find any item from this list that accept this contextMenuEvent.
In our case the proxy will always accept it and therefor will always send
it to the widget that it owns and that's why we had this infinite numbers
of QMenu. In QWidget world the context menu event is always sent to the
widget that has the focus. If you checked any QWidget subclasses you'll
see that all of them assume that when they get a context menu event
everything is fine focus wise. We could have changed
QGraphicsScene::contextMenuEvent but this will breaks some existing code.
Instead we just checked that QGraphicsProxyWidget has the focus
before we actually send the contextMenuEvent to the widget it owns.

I have modified an existing auto-test to cover that. Be careful
widget->setContextMenuPolicy(Qt::ActionsContextMenu) means that you will
get no contextMenuEvent. :D. In this test i cover the standard use case
and the one with Qt::ActionsContextMenu with and without the focus on the
proxy.

Task-number:QTBUG-3787
Reviewed-by:jan-arve