QGraphicsAnchorLayout: Add test for 'infinite' max size widgets
With four items anchored side by side, each of them being allowed
to grow to QWIDGETSIZE_MAX, we have a situation of fair distribution
because the layout itself can grow only to QWIDGETSIZE_MAX (and not
four times this amount).

This test identified an error on the expanding distribution logic.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>