Fix .ui of Styledemo (Layout tweaks)
Let's make sure to have enough layout spacings around
the widgets in order to see the focusrect.
Also, the issue with the growing dialog is fixed. It
was the spinbox and it's label which with the stylesheets
set did not have enough horizontal space, and thus
made the dialog grow on the following CSS change

Reviewed-By: TrustMe