Symbian fix: rename examples ftp and musicplayer to start with q.
Symbian already has the executables musicplayer.exe, mediaplayer.exe and
ftp.exe, with the result that we overwrite them with Qt. We solve this
by renaming the examples, and do it not only on Symbian, such that Qt
remains consistent across platforms.

This was previously partly done for mediaplayer. The folder name needs
to be consistent with the executable, for qtdemo to work. Done jointly
with Alessandro.

Task-number: QTBUG-4743

Reviewed-by: Alessandro Portale
Reviewed-by: Frans Englich