QUuid::createUuid() not unique when using threads on Unix
QUuid::createUuid() only seeds the PRNG on the first entry, but since
it's using qsrand() and qrand(), all other threads will use the
default seed, and thus generate the exact same UUIDs.

Fix this by adding an internal function (qsrand() overload with no
args) which seeds the PRNG if it hasn't been done already, and use a
seed that is based on current time, a stack address and a global
serial counter (so that the chances of 2 threads using the same seed
are as low as possible).

Task-number: QTBUG-3543
Reviewed-by: Marius Storm-Olsen