Fixes QPauseAnimation autotests for unreliable timer intervals on Windows
The timer interval used currently on Windows is 16 ms, but we get ticks
at every 32 ms on average, so the consistent timing is not reliable on windows.
We should use the multimedia timer instead (use 15 ms for QTimer), once
qt is able to handle events while native event loops are running.

When this is done, the ifdefs introduced in this commit should be removed.

Reviewed-by: thierry