Fix printing bitmap fonts on X11 with FontConfig enabled
When FontConfig was enabled, bitmap fonts would often get a different
pixel size than the one we requested. Usually the size would only be
a pixel off, but this was especially visible when printing in highres
with bitmap fonts, because in those cases we would request a pixel size
which was computed based on the printer's high dpi.

The result was that all printed text with bitmap fonts would be really
really tiny. The fix falls back to using the XLFD font engine when
using bitmap fonts (when the returned pixel size is different from the
requested), because this engine scales the fonts for us. This will
cause bitmap fonts to be rendered without antialiasing.

Task-number: QTBUG-3620
Reviewed-by: Simon Hausmann