Fixed enum values in Qt::GestureContext.
The only reason to refer to values from the Qt::ShortcutContext enum is
to avoid confusing people since enum value names look similar so we want
to avoid weird behaviour when mixing them. But referring to another enum
value makes the documentation look weird as it mentions different
unrelated enum value in the GestureContext doc.

Reviewed-by: trustme