images: Document QImageReader::loopCount behaviour for infinite loops
The gif and mng reader will return -1 as QImageReader::loopCount()
for infinite images. Document that this -1 is clashing with the
error handling of this method.

The documentation was mixing the default and error cases. In some
cases "Otherwise -1" and "Otherwise 0" was used. Change this to
mention the default coming from QImageIOHandler in the "Otherwise"
clause and document the error case with -1.

Add a test case to check that loopCount() returns -1 for infinite
animations.

Reviewed-by: Andreas