Changes to support locale change event for symbian platform
Subscribing to the locale change notification to be able to update the
system locale whenever the user changes the current system locale.

Also changed the initialization of the system locale to make construction of
the QLocale object as lightweight as possible. So now the default contructor
just creates a QLocale and QSystemLocale objects, but doesn't try to fill the
cache in the latter with data from the system and postpones it until it is
actually requested (most applications create QLocale objects on the stack and
might not even use the data from the system locale, so we don't need to
initialize system locale right away).

Modified-by: axis
Modified-by: Denis Dzyubenko

Reviewed-by: Denis Dzyubenko
Reviewed-by: axis