gif: 10% speedup on microbenchmark on QImageReader
On the microbenchmark of what is executed during
pure pageloading in regarding to QImageReader this
is an improvement.

Do not use the non const QImage::scanLine as this will
go through the QImage::detach all the time making
QImage::scanLine show up in the profile.

Introduce a FAST_SCAN_LINE that is doing the scanLine
calculation on local variables.

Reviewed-by: Andreas