Re-apply change 6b8ac349b9a477863a8c8388dcc0658f3284bc54 by Jocelyn Turcotte
Re-applying commit ee0a43fee20cc398b505eb65218ebed56dfc8f39
by Simon Hausmann

Fix crash of QtScript on Mac OS X

When compiling on 10.4 but running on 10.5 the flags passed to vm_map cause
it to crash.

For now fall back to the use of mmap() as allocator instead.

Reviewed-by: Kent Hansen