Fixed wrong use of graphics effects for pixmap graphics items.
The blur, drop shadow, and bloom graphics effects are scale dependent,
since they have radius and offset (in the case of drop shadow)
parameters that are specified in device coordinates. Thus, we can't
apply the effect in logical coordinates and scale up, and need to always
use the device coordinate path for these effects. The opacity and
grayscale effects still use the logical coordinate optimization.

Reviewed-by: Gunnar Sletta