Sorting bug in QTableView fix
When the DisplayRole is identical for two or more items, any
modification of one item (e.g. backgroundColorRole) may trigger a
re-ordering, which is obviously not an behaviour to be expected.
Same fix as the one used for QTreewidget in
c9eacfa1c791e2d228a3c8f0c119d02d7f46ee02.

Task-number: QTBUG-4856
Task-number: 262056
Reviewed-by: Olivier Goffart