QGAL: refactor and document the simplification algorithm
Refactor the simplifyGraphIteration() function. The aim was to make it
more clear without changing too much the way it does stuff.

Before it collected a list of candidates and then filtered that into
sublists of same order, and after that removed the center edges from
the borders. It also reversed the list if the direction wasn't
forward, to pass it in forward order to simplifySequentialChunk()
helper function.

The refactored version

- take in account the order when building the candidates, this avoids
index manipulation;
- try to calculate 'beforeSequence' and 'afterSequence' as it builds the
candidates list;
- make simplifySequentialChunk() aware of directions, now it deals internally
with reversed direction sequences.

This commits also adds explanations to trickier parts of the code.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>