Fixed QPainter::begin() so that it fails gracefully.
Fixed QPainter::begin() so that if it fails, it can still be used on
other paint devices without causing an assert. Autotest included.
I also made begin() fail on images with the QImage::Format_Indexed8
format and added warnings in the documentation since painting on such
images is not supported.

Reviewed-by: Trond