QGAL: Add anchor layout vs nested linear layouts benchmark
Adding a benchmark to compare the use of three nested linear layouts
versus the use of a single anchor layout.

We also test two different anchor layout setups to achieve the
same visual result. In the first one we use the addCornerAnchors API
that, in the way it was used, adds reduntant anchors. The second uses
the basic addAnchor API to create only the essential anchors.

Currently the first setup is way slower than the second. That happens
because the redundant anchors create "zero-sized knots" in the graph,
that are not eaten by the simplification algorithm, thus requiring
the usage of the simplex solver.

Zero-sized knots are groups of three or more vertices that are linked
together by zero-sized anchors. In pratice, these vertices represent
the same place in the graph (remember, the distance between them is
zero), but to the simplification mechanism, they look like three
distinct ones.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>