Softkeys remain disabled if action owning action widget is enabled
Softkeys have a QAction that is related to some action widget.
The initial state of the action was set according the state of action
widget (enabled/disabled). Now, if action widget's state changes,
the softkey's action remain in the initial state.

This was fixed by removing the enable/disable from the QAction and
instead use the real state of action widget when handling the
command of softkey.

Task-number: QTBUG-4619
Reviewed-by: Janne Anttila