Send posted events in response to WM_QT_SENDPOSTEDEVENTS
(which is just WM_USER+1)

Delay the next WM_QT_SENDPOSTEDEVENTS iff there is a WM_TIMER
or input event pending

We also need to break out of processEvents() after seeing this
message, to prevent livelocking in the prescence of fast timers.

I also took the liberty of defining WM_QT_SOCKETNOTIFIER (WM_USER)
at the same time (to give clear meaning to what WM_USER and WM_USER+1
are used for).

Reviewed-by: Prasanth Ullattil