Fix autotest case for QSoftkeyManager
QSoftkeyManager's test case checkSoftkeyEnableStates was broken with
recent fix to QSoftkeyManager where softkey action is no longer
initialized to the initial state of action widget. Instead, softkey
action checks the state of action widget when handling the action.

Autotest case is now fixed to handle the change.

Task-number: N/A
Reviewed-by: Miikka Heikkinen