QGAL: identify unfeasible setups even when graph is simplified
The calculate graphs now can return early due to unfeasible anchor
setups found out during simplification. This allows finding out
problems in parallel anchors, when one anchor has maximum size smaller
than the minimum size of the other anchor.

The order for simplification and refreshing size hints also has been swaped:

- First, refresh size hints for all anchors in the graph. If the graph is
  simplified, the refreshSizeHints() call will reach the children anchors.
- Then, if the simplificated graph is invalid, rebuild it. During this
  rebuild, refreshSizeHints_helper() will be called for all levels.

So in both situations we can identify an unfeasible setup.  Note that
this test alone is not enough to classify the graph as feasible, depending
on the graph, it will still need to go through the Simplex.

A test case was added and the function that traverse the graph
refreshing now is called refreshAllSizeHints(). The old name was not
so clear since the function will not fill only the anchors that have
items associated.

Last but not least, the lastCalculationUsedSimplex variable is cleared
when starting calculateGraphs(), since we now can leave the function
earlier, without reaching calculateTrunk(), which is the function that
sets it.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>