Optimize QVGPixmapDropShadowFilter by removing colorize step
Previously, the drop shadow was colorizing the incoming
image and then blurring the colorized version.  This change
first blurs the image to an alpha-only VGImage and then uses
that VGImage as a stencil to draw the drop shadow color.

This way, there is only 1 filter step and a draw instead of
2 filter steps and a draw.  The result is to make the performance
of the drop shadow filter almost identical to the blur filter.

Reviewed-by: trustme