Mac: fix bug when reparenting non-native menu bars
If we reparent a naive menu bar inside a widget that
has a menu bar from before, the old menu bar is lost.
This patch makes sure that we don't set the new
menu bar in this case, but let it remain as-is in
case the widget is reparanted out again later.

Rev-By: MortenS