QToolButton popup menu is shown at wrong position when embedded in a QGraphicsView.
The main problem here is that QWidget assume that they are in the screen
somewhere, which means inside the available geometry provided by
QDesktopWidget. But in QGraphicsView the button can be in a position
that is way bigger than the screen resolution. Lot of widgets make
this assumption when positionning subpopups or submenus. Instead of
applying the same code on tons of QWidgets, it's better to have an
helper function in desktop widget which catch this case. It's not
pretty (since it has nothing to do with QDesktopWidget) but we don't
have better solution.

Task-number:QTBUG-3822
Reviewed-by:brad