Fix bug in embedded dialog demo with tab focus.
On embedded dialog pressing tab stop changing the focus when the focus
was given to QFontComboBox. It's because QFontComboBox embed a
QLineEdit in order to allow editing. But this QLineEdit is a focus proxy
so we need to special case that. The logic is the same in QApplication.
Be careful when changing one of them.

Task-number:QTBUG-4818
Reviewed-by:jan-arve
Reviewed-by:ogoffart