Improved QFontInfo::pointSize() slightly on X11.
In non-GUI applications on X11, QFont and QFontInfo return different
point size because for QFontInfo, the point size is converted to pixel
size and back, but with different dpis. This commit improves the
situation for the case where font config is used, but the bug still
needs to be fixed properly by using the same dpi for all point<->pixel
size conversions.

Reviewed-by: Trond