Properly detect font smoothing setting on Mac OS X in raster engine
We would assume font smoothing on the mac was always turned on, giving
poor text rendering in the cases where it was not. This implementation
mirrors querying the cleartype setting on Windows, checking the setting
on application initialization and rendering into an 8 bit cache if it
is turned off.

Task-number: QTBUG-4881
Reviewed-by: Morten Johan Sørvig