Make Symbian emulator behave the same as device for mouse & touchscreen
Emulator HAL reports that it supports mouse but not touch.
Since Qt is changing behaviour based on these values, the emulator works
differently from the device.

To workaround, the emulator HAL values are ignored and following used:
5.0: mouse 0, touch 1
3.x: mouse 0, touch 0

The mouse can still be used to interact with the emulator, as pointer
events are still delivered.

Task-number: QTBUG-4803
Task-number: QTBUG-4616
Reviewed-by: Sami Merila