Video screen region is updated in response to ancestors of video widget being moved.
Because QWidget::moveEvent is only called when a widget moves relative
to its parent, a widget's absolute screen position may change without it
receiving a moveEvent (for example, as a result of its parent being
moved).

The MMF video playback API on Symbian v9.4 requires, in addition to a
window handle, an absolute screen rectangle.  Changes in the video
widget's absolute screen position therefore need to be propagated
into the MMF.

This change introduces a new object, AncestorMoveMonitor, which
installs an event filter on the QCoreApplication instance.  A
VideoOutput object registers with the AncestorMoveMonitor, which
listens on its behalf for MoveEvents and ParentChangeEvents directed
at any of the ancestors of the VideoOutput.  MoveEvents trigger a
callback to the VideoOutput instance, which then notifies the MMF of
the new screen rectangle.  ParentChangeEvents cause the
AncestorMoveMonitor to update the ancestor list associated with the
target VideoOutput instance.

The video position now tracks that of the associated widget, but there
are two problems which require further investigation:

1. The video window lags behind.  This may be an unavoidable
consequence of the fact that setting a new screen rectangle causes the
MMF to tear down its DSA session and start a new one; this is known to
block the window server and take some time to complete.

2. Artifacts are visible around the edges of the moving video widget.

Task-number: QTBUG-4787
Reviewed-by: Frans Englich