Honor graphics system on Mac/Cocoa when exposing and resizing window
When exposing or resizing a previously hidden window on Cocoa, we would
go through the CoreGraphicsPaintEngine even when a different graphics
system was set. This was because there were two different paths from
the windowing system into our paint event. The one going through the
virtual drawRect function in QCocoaView did not honor the graphics
system. This patch makes sure the backing store is used for these types
of events as well.

Done with: Gunnar

Reviewed-by: MortenS