Fix Qt containers to properly support types with strict alignments.
QContiguousCache is a new type, so there are no binary compatibility
issues.

QHash and QMap didn't have any public qMalloc / qFree, so the entire
logic is contained in .cpp code. However, since old code will not
inform us of the alignment requirements, we need to add a bit to the
structure to indicate whether strict alignment is in use or not.

QList doesn't require any changes. For small, movable types, they're
all stored in the pointer array itself, so they're aligned. For larger
types, we use new(), so types with stricter requirements should define
their own operator new().

QLinkedList cannot be fixed. It uses new() on the QLinkedListNode,
which contains a T type. Sorry.

QVector did have public qMalloc / qFree. I've moved the calls to the
inner function and made it keep the old calls if the alignment
requirement is below a certain threshold. The idea is that, if it's
above, no one was using QVector anyway.

Reviewed-by: Bradley T. Hughes