Avoid calls to public QScriptValue::engine() function
Calls to engine() are mostly done for checking that the "source"
and "target" engines are the same, but we don't want those checks
to slow us down. Use an inline getEngine() function instead.
This makes e.g. QScriptValue::call() ~10% faster for a function
like "function(a, b) { return a + b; }".

Reviewed-by: Olivier Goffart