Added visual DnD indicator for left to right flow in QListView
Extended the behaviour already present in QAbstractItemView into
QListView. This means some code duplication which should maybe be refactored
at some point. (Seems to be the price to pay when some delegate does almost
the same thing as the base class).

No auto-test, as it's a drag-and-drop related task.

Reviewed-by: Thierry
Task-number: QTBUG-3730