Removed pixelize, bloom and grayscale filter
This is new API and we don't want to add several items that are not
strictly needed. This is a new set of features and we can grow them
once we have more input from users on what is needed.

The Bloom filter was added based on input from designers, but is not
implemented according to how designers think of blook, so the effect
doesn't meet the requirements.

The Grayscale filter is functionally a duplicate of the colorize
filter and is therefore not needed.

The Pixelize filter has no genuine usecase.

Reviewed-by: Samuel