On Windows, report a 0 file size for streams and other funny files
Obtain file size directly, instead of relying on fseek/ftell and messing
with the file position.

Also changed the return value on errors to 0. This is mostly relevant on
streams and pipes, where we used to return whatever error value ftell
returned (usually -1). This change also makes the return value
consistent with what is returned on Unix platforms and what we document
for Windows CE.

Nevertheless, documentation of this and related issues is misleading and
needs to be updated.

Reviewed-by: Markus Goetz