Wrong worldTransform() on the painter in QGraphicsScene::drawForeground.
The painter's worldTransform() is updated for each item we draw, and
when the DontSavePainterState optimization flag is set, this change is
not protected by save() and restore(). After all the items are drawn, it
means the painter is left with the last drawn item's transform. We
therefore have to make sure it is reset back to whatever it was before
the items were drawn.

Auto-test included.

Task-number: QTBUG-4973
Reviewed-by: alexis
Reviewed-by: andreas