Fix QFile::isSequential on Windows
When not using native HANDLEs, the return of isSequential was hardcoded
to true for files with a fd, and for the standard FILE* streams stdin,
stdout and stderr; false for all other FILE* streams.

We now use the native GetFileType call for all files by obtaining a
native handle where required. We also treat files of type FILE_TYPE_CHAR
as sequential, as is the case for the standard streams in console
applications.

When standard streams are redirected to/from files, GetFileType will
return FILE_TYPE_DISK for them and they won't be considered sequential.
This is alright since in this mode they behave like regular files and
QFile::seek() will work for random offsets.

Reviewed-by: Marius Storm-Olsen