Add cross compiling pkg-config disabled warning
We have historically silently disabled pkg-config when cross compiling in
order to avoid host pkg-config contamination of the target build.
pkg-config files are present on many embedded targets, and make the correct
detection/configuration of dependencies far more convenient for the
developer. It is therefor valuable to advertise that pkg-config
functionality is available but disabled barring direct involvement on their
parts, rather than silently disabling it.

Reviewed-by: Thiago Macieira <thiago.macieira@nokia.com>