Autotest: fix improper use of the serviceOwnerChanged signal
This test was doubly wrong: it first registered a service name, then
it connected to signal to watch it. You can't receive a signal if you
connect to it after it's emitted...

Second, it waited for any serviceOwnerChanged() signal to exit the
event loop, not necessarily the one we wanted to receive.

This used to work because we'd always connect to the D-Bus signal, but
now we don't anymore.