Add an extra overload for QMatrix4x4::toTransform()
Change 100afe8d fixed a bug in QGraphicsRotation related to when
the "distance to plane" projection needed to be performed.  As a
side effect it made the toTransform() API not do the expected
thing when the function is called with no argument.

This change makes the default no-argument version of toTransform()
do the simple "drop row 3 and column 3" orthographic transformation
that normal users of the class expect, and adds a new overload for
the "distance to plane" projection case for the special case.

Reviewed-by: trustme